image_editing/requirements.txt
Pillow==9.1.*